Extract PosixModeBits.ts into a separate source file
